feat: split hub into inference center + marketplace
1. Left sidebar brand: Jan -> RongxinAI
2. Sidebar nav: separate '推理中心' (/hub/) and '模型市场' (/marketplace/)
   - 推理中心: Ollama status, install/start, placeholder for future params
   - 模型市场: full ModelScope model catalog with search, sort, download
3. Ollama start feedback: toast loading/success/warning with multi-poll
4. NSIS installer: SimpChinese localization + translated ollama.nsh
5. Add modelMarketplace i18n key to all 15 locales
