+++ src-tauri/nsis/SimpChinese.nsh
@@ -1,0 +1,29 @@
+LangString addOrReinstall ${LANG_SIMPCHINESE} "添加/重新安装组件"
+LangString alreadyInstalled ${LANG_SIMPCHINESE} "已安装"
+LangString alreadyInstalledLong ${LANG_SIMPCHINESE} "${PRODUCTNAME} ${VERSION} 已安装。选择要执行的操作，然后点击下一步继续。"
+LangString appRunning ${LANG_SIMPCHINESE} "{{product_name}} 正在运行！请先关闭它，然后重试。"
+LangString appRunningOkKill ${LANG_SIMPCHINESE} "{{product_name}} 正在运行！$
+点击确定以关闭它"
+LangString chooseMaintenanceOption ${LANG_SIMPCHINESE} "选择要执行的维护选项。"
+LangString choowHowToInstall ${LANG_SIMPCHINESE} "选择您要安装 ${PRODUCTNAME} 的方式。"
+LangString createDesktop ${LANG_SIMPCHINESE} "创建桌面快捷方式"
+LangString dontUninstall ${LANG_SIMPCHINESE} "不要卸载"
+LangString dontUninstallDowngrade ${LANG_SIMPCHINESE} "不要卸载（此安装程序禁用了未卸载直接降级）"
+LangString failedToKillApp ${LANG_SIMPCHINESE} "无法关闭 {{product_name}}。请先关闭它，然后重试"
+LangString installingWebview2 ${LANG_SIMPCHINESE} "正在安装 WebView2..."
+LangString newerVersionInstalled ${LANG_SIMPCHINESE} "已安装更新版本的 ${PRODUCTNAME}！不建议安装旧版本。如果确实要安装此旧版本，建议先卸载当前版本。选择要执行的操作，然后点击下一步继续。"
+LangString older ${LANG_SIMPCHINESE} "旧版"
+LangString olderOrUnknownVersionInstalled ${LANG_SIMPCHINESE} "您的系统上已安装 $R4 版本的 ${PRODUCTNAME}。建议在安装前先卸载当前版本。选择要执行的操作，然后点击下一步继续。"
+LangString silentDowngrades ${LANG_SIMPCHINESE} "此安装程序禁用了降级，无法使用静默安装程序继续，请改用图形界面安装程序。$
+"
+LangString unableToUninstall ${LANG_SIMPCHINESE} "无法卸载！"
+LangString uninstallApp ${LANG_SIMPCHINESE} "卸载 ${PRODUCTNAME}"
+LangString uninstallBeforeInstalling ${LANG_SIMPCHINESE} "安装前先卸载"
+LangString unknown ${LANG_SIMPCHINESE} "未知"
+LangString webview2AbortError ${LANG_SIMPCHINESE} "安装 WebView2 失败！没有 WebView2 应用无法运行。请尝试重新启动安装程序。"
+LangString webview2DownloadError ${LANG_SIMPCHINESE} "错误：下载 WebView2 失败 - $0"
+LangString webview2DownloadSuccess ${LANG_SIMPCHINESE} "WebView2 引导程序下载成功"
+LangString webview2Downloading ${LANG_SIMPCHINESE} "正在下载 WebView2 引导程序..."
+LangString webview2InstallError ${LANG_SIMPCHINESE} "错误：安装 WebView2 失败，退出代码 $1"
+LangString webview2InstallSuccess ${LANG_SIMPCHINESE} "WebView2 安装成功"
+LangString deleteAppData ${LANG_SIMPCHINESE} "删除应用程序数据"
