+++ src-tauri/nsis/ollama.nsh
@@ -11,7 +11,7 @@
   ${EndIf}
 
   ${If} $R0 != ""
-    DetailPrint "Ollama already installed ($R0), skipping"
+    DetailPrint "检测到 Ollama 已安装 ($R0)，跳过安装"
     Goto ollama_done
   ${EndIf}
 
@@ -23,29 +23,29 @@
   ; 3. Prefer local pre-staged installer if it exists
   ${If} $R1 != ""
   ${AndIf} ${FileExists} $R1
-    DetailPrint "Using local OllamaSetup.exe: $R1"
+    DetailPrint "使用本地 OllamaSetup.exe: $R1"
     StrCpy $R0 $R1
     Goto do_ollama_install
   ${EndIf}
 
   ; 4. Otherwise download from configured URL
   ${If} $R2 == ""
-    DetailPrint "No local OllamaSetup.exe found and no DownloadUrl configured"
+    DetailPrint "未找到本地 OllamaSetup.exe，且未配置下载地址"
     Goto ollama_done
   ${EndIf}
 
-  DetailPrint "Downloading Ollama from $R2 ..."
+  DetailPrint "正在从 $R2 下载 Ollama ..."
   Delete "$TEMP\OllamaSetup.exe"
   NSISdl::download $R2 "$TEMP\OllamaSetup.exe"
   Pop $R0
   ${If} $R0 != "success"
-    DetailPrint "Ollama download failed: $R0"
+    DetailPrint "Ollama 下载失败: $R0"
     ${IfNot} ${Silent}
-      MessageBox MB_ICONEXCLAMATION|MB_YESNO "Failed to download Ollama. Continue without Ollama?" IDYES continue_without_ollama_dl
-      Abort "Installation cancelled due to Ollama download failure"
+      MessageBox MB_ICONEXCLAMATION|MB_YESNO "Ollama 下载失败。是否继续安装（不安装 Ollama）？" IDYES continue_without_ollama_dl
+      Abort "因 Ollama 下载失败，安装已取消"
       continue_without_ollama_dl:
     ${Else}
-      DetailPrint "Silent mode: continuing without Ollama"
+      DetailPrint "静默模式：跳过 Ollama，继续安装"
     ${EndIf}
     Goto ollama_done
   ${EndIf}
@@ -54,10 +54,10 @@
 do_ollama_install:
   ; Build command line (silent + optional custom directory)
   ${If} $R3 != ""
-    DetailPrint "Installing Ollama silently to $R3 ..."
+    DetailPrint "正在静默安装 Ollama 到 $R3 ..."
     ExecWait '"$R0" /S /DIR="$R3"' $R1
   ${Else}
-    DetailPrint "Installing Ollama silently with default directory ..."
+    DetailPrint "正在静默安装 Ollama（使用默认目录）..."
     ExecWait '"$R0" /S' $R1
   ${EndIf}
 
@@ -67,15 +67,15 @@ do_ollama_install:
   ${EndIf}
 
   ${If} $R1 = 0
-    DetailPrint "Ollama installed successfully"
+    DetailPrint "Ollama 安装成功"
   ${Else}
-    DetailPrint "Ollama installation failed with exit code: $R1"
+    DetailPrint "Ollama 安装失败，退出代码: $R1"
     ${IfNot} ${Silent}
-      MessageBox MB_ICONEXCLAMATION|MB_YESNO "Ollama installation failed. Continue without Ollama?" IDYES continue_without_ollama
-      Abort "Installation cancelled due to Ollama installation failure"
+      MessageBox MB_ICONEXCLAMATION|MB_YESNO "Ollama 安装失败。是否继续安装（不安装 Ollama）？" IDYES continue_without_ollama
+      Abort "因 Ollama 安装失败，安装已取消"
       continue_without_ollama:
     ${Else}
-      DetailPrint "Silent mode: continuing without Ollama"
+      DetailPrint "静默模式：跳过 Ollama，继续安装"
     ${EndIf}
   ${EndIf}
