+++ web-app/src/components/left-sidebar/index.tsx
@@ -20,7 +20,7 @@ export function LeftSidebar() {
       <Sidebar variant="floating" collapsible="offcanvas">
         <SidebarHeader className="flex px-1">
           <div className={cn("flex items-center w-full justify-between", IS_MACOS && "justify-end")}>
-            {!IS_MACOS && <span className="ml-2 font-medium font-studio">Jan</span>}
+            {!IS_MACOS && <span className="ml-2 font-medium font-studio">RongxinAI</span>}
             <div>  
               {isLeftPanelOpen && <DownloadManagement />}
               <SidebarTrigger className="text-muted-foreground rounded-full hover:bg-sidebar-foreground/8! -mt-0.5 relative z-50 ml-0.5" />
