+++ web-app/src/locales/pl/common.json
@@ -397,5 +397,6 @@
   "conversationNotAvailableDescription": "Rozmowa, do której próbujesz uzyskać dostęp, nie jest dostępna lub została usunięta.",
   "temporaryChat": "Tymczasowy Czat",
   "temporaryChatTooltip": "Tymczasowy czat nie pojawi się w Twojej historii",
-  "inferenceCenter": "推理中心"
+  "inferenceCenter": "推理中心",
+  "modelMarketplace": "Rynek Modeli"
 }
