+++ web-app/src/locales/zh-CN/common.json
@@ -397,5 +397,6 @@
       "description": "线程已成功从\"{{projectName}}\"中移除"
     }
   },
-  "inferenceCenter": "推理中心"
+  "inferenceCenter": "推理中心",
+  "modelMarketplace": "模型市场"
 }
